Move: OmosApplication 경로 수정
