+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
     <application
-        android:name=".di.OmosApplication"
+        android:name=".OmosApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher_omos"
         android:label="@string/app_name"
@@ -39,7 +39,20 @@
 
         </activity>
 
+        <activity
+            android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
 
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <data
+                    android:host="oauth"
+                    android:scheme="kakao${KAKAO_OAUTH_KEY}" />
+            </intent-filter>
+        </activity>
 
 
         <activity
@@ -151,26 +164,10 @@
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/MainStatusBarStyle" />
-        <activity
-            android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
 
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.BROWSABLE" />
-
-                <!-- Redirect URI: "kakao{NATIVE_APP_KEY}://oauth" -->
-                <data
-                    android:host="oauth"
-                    android:scheme="kakao${KAKAO_OAUTH_KEY}" />
-            </intent-filter>
-        </activity>
         <activity
             android:name=".ui.onboarding.LoginActivity"
             android:screenOrientation="portrait" />
-
-
 
         <!-- 외부 저장소 사용을 위한 FileProvider auth 추가. -->
         <provider
