+++ app/src/main/java/com/infinity/omos/di/NetworkModule.kt
@@ -104,7 +104,7 @@ object NetworkModule {
         converterFactory: Converter.Factory,
     ): Retrofit {
         return Retrofit.Builder()
-            .baseUrl(BuildConfig.BASE_URL)
+            .baseUrl(BuildConfig.OMOS_BASE_URL)
             .client(okHttpClient)
             .addConverterFactory(converterFactory)
             .build()
