+++ app/src/main/java/com/infinity/omos/repository/ReportBlockRepository.kt
@@ -7,7 +7,7 @@ import com.infinity.omos.api.ReportService
 import com.infinity.omos.api.RetrofitAPI
 import com.infinity.omos.data.ReportBlock
 import com.infinity.omos.data.ResultState
-import com.infinity.omos.di.OmosApplication
+import com.infinity.omos.OmosApplication
 import com.infinity.omos.utils.NetworkUtil
 import retrofit2.Call
 import retrofit2.Callback
