+++ app/src/main/java/com/infinity/omos/repository/SearchRepository.kt
@@ -7,7 +7,7 @@ import com.infinity.omos.api.RetrofitAPI
 import com.infinity.omos.api.SearchService
 import com.infinity.omos.data.*
 import com.infinity.omos.etc.Constant
-import com.infinity.omos.di.OmosApplication
+import com.infinity.omos.OmosApplication
 import com.infinity.omos.utils.NetworkUtil
 import retrofit2.Call
 import retrofit2.Callback
