+++ app/src/main/java/com/infinity/omos/ui/main/MyDjFragment.kt
@@ -21,7 +21,7 @@ import com.infinity.omos.adapters.MyDjListAdapter
 import com.infinity.omos.databinding.FragmentMyDjBinding
 import com.infinity.omos.etc.Constant
 import com.infinity.omos.utils.CustomDialog
-import com.infinity.omos.di.OmosApplication
+import com.infinity.omos.OmosApplication
 import com.infinity.omos.viewmodels.SharedViewModel
 
 class MyDjFragment : Fragment() {
