+++ app/src/main/java/com/infinity/omos/ui/main/MyPageFragment.kt
@@ -19,7 +19,7 @@ import com.infinity.omos.databinding.FragmentMyPageBinding
 import com.infinity.omos.etc.GlobalFunction
 import com.infinity.omos.ui.dj.FollowActivity
 import com.infinity.omos.ui.setting.SettingActivity
-import com.infinity.omos.di.OmosApplication
+import com.infinity.omos.OmosApplication
 import com.infinity.omos.viewmodels.SharedViewModel
 
 class MyPageFragment : Fragment() {
