+++ app/src/main/java/com/infinity/omos/ui/main/MyRecordFragment.kt
@@ -17,7 +17,7 @@ import com.infinity.omos.R
 import com.infinity.omos.adapters.MyRecordListAdapter
 import com.infinity.omos.databinding.FragmentMyRecordBinding
 import com.infinity.omos.etc.Constant
-import com.infinity.omos.di.OmosApplication
+import com.infinity.omos.OmosApplication
 import com.infinity.omos.viewmodels.SharedViewModel
 
 class MyRecordFragment : Fragment() {
