+++ app/src/main/java/com/infinity/omos/ui/write/WriteRecordActivity.kt
@@ -29,7 +29,7 @@ import com.infinity.omos.databinding.ActivityWriteRecordBinding
 import com.infinity.omos.etc.GlobalFunction
 import com.infinity.omos.utils.AWSConnector
 import com.infinity.omos.utils.CustomDialog
-import com.infinity.omos.di.OmosApplication
+import com.infinity.omos.OmosApplication
 import com.infinity.omos.viewmodels.WriteRecordViewModel
 import com.theartofdev.edmodo.cropper.CropImage
 import com.theartofdev.edmodo.cropper.CropImageView
